2.1.0: Integrating monitoring. git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@2343 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/producer/transition/transition_producer.h
@@ -64,7 +64,7 @@ struct transition_info
 		, direction(transition_direction::from_left)
 		, tweener(L"linear"){}
 		
-	size_t					duration;
+	int						duration;
 	transition_direction	direction;
 	transition_type			type;
 	tweener					tweener;
